split bsp conversion stuff from q3map2
