+++ tools/quake3/q3map2/q3map2.vcproj
@@ -319,6 +319,10 @@
 				>
 			</File>
 			<File
+				RelativePath=".\convert_bsp.c"
+				>
+			</File>
+			<File
 				RelativePath=".\convert_map.c"
 				>
 			</File>
